requirement.txt
google
googlesearch-python
beautifulsoup4
requests
pyfiglet
